fix: 데미지 매개변수 타입 수정

+++ Assets/Dev/Hak/Interaction/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void Damage(int damage);
+    public void Damage(float damage);
 }
